requirements.txt
apturl==0.5.2
asgiref==3.5.2
backports.zoneinfo==0.2.1
bcrypt==3.1.7
Beaker==1.11.0
blinker==1.4
Brlapi==0.7.0
certifi==2019.11.28
chardet==3.0.4
charset-normalizer==2.1.1
Click==7.0
colorama==0.4.3
command-not-found==0.3
cryptography==2.8
cupshelpers==1.0
dbus-python==1.2.16
decouple==0.0.7
defer==1.0.6
defusedxml==0.7.1
distlib==0.3.5
distro==1.4.0
distro-info===0.23ubuntu1
Django==4.1.2
django-cors-headers==3.10.1
django-crispy-forms==1.14.0
django-filter==22.1
django-forestadmin==1.4.2
django-ipware==4.0.2
djangorestframework==3.14.0
duplicity==0.8.12.0
ecdsa==0.18.0
entrypoints==0.3
fasteners==0.14.1
filelock==3.7.1
future==0.18.2
httplib2==0.14.0
idna==2.8
importlib-metadata==5.0.0
keyring==18.0.1
language-selector==0.1
launchpadlib==1.10.13
lazr.restfulclient==0.14.2
lazr.uri==1.0.3
lockfile==0.12.2
louis==3.12.0
macaroonbakery==1.3.1
Mako==1.1.0
Markdown==3.4.1
MarkupSafe==1.1.0
marshmallow==3.14.1
marshmallow-jsonapi==0.24.0
monotonic==1.5
netifaces==0.10.4
numpy==1.23.4
oauthlib==3.1.0
oic==1.4.0
olefile==0.46
pandas==1.5.0
paramiko==2.6.0
pbr==5.11.0
pexpect==4.6.0
Pillow==7.0.0
platformdirs==2.5.2
protobuf==3.6.1
pyasn1==0.4.8
pycairo==1.16.2
pycryptodomex==3.15.0
pycups==1.9.73
PyGObject==3.36.0
pyjwkest==1.4.2
PyJWT==1.7.1
pymacaroons==0.13.0
PyNaCl==1.3.0
pyRFC3339==1.1
python-apt==2.0.0+ubuntu0.20.4.7
python-dateutil==2.8.2
python-debian===0.1.36ubuntu1
python-jose==3.3.0
pytz==2022.4
pyxdg==0.26
PyYAML==5.3.1
reportlab==3.5.34
requests==2.28.1
requests-unixsocket==0.2.0
rsa==4.9
SecretStorage==2.3.1
simplejson==3.16.0
six==1.14.0
sqlparse==0.4.3
systemd-python==234
testresources==2.0.1
typing_extensions==4.4.0
ubuntu-advantage-tools==27.8
ubuntu-drivers-common==0.0.0
ufw==0.36
unattended-upgrades==0.1
Unipath==1.1
urllib3==1.25.8
usb-creator==0.3.7
virtualenv==20.16.0
wadllib==1.3.3
xkit==0.0.0
zipp==3.9.0
